# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "HospitalPatient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "condition": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "admittedAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/patients/search": {
      "get": {
        "tags": [
          "hospitals"
        ],
        "summary": "Busca pacientes hospitalizados por nombre u otros datos",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Texto de búsqueda"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Máximo de resultados (se acota entre 1 y 200)"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de la búsqueda de pacientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HospitalPatient"
                      }
                    },
                    "query": {
                      "type": "string"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas peticiones (rate limit)"
          }
        }
      }
    },
...
    "/api/reports/{id}/confirm": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Confirma un reporte de emergencia (una vez por IP)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmación registrada; devuelve el total de confirmaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "confirmations": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "La IP ya confirmó este reporte",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas confirmaciones desde el dispositivo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo confirmar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Lista de reportes de emergencia",
        "responses": {
          "200": {
            "description": "Reportes y bandera de persistencia"
          }
        }
      },
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Crear un reporte de emergencia",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lat",
                  "lng",
                  "place",
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "place": {
                    "type": "string"
                  },
                  "affected": {
                    "type": "integer"
                  },
                  "needs": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "description": "data:image/...;base64 (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reporte creado"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "No se pudo guardar"
          }
        }
      }
    },
...
    "/api/sync/cron": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Dispara el cron de sincronización de fuentes externas (Auth Bearer CRON_SECRET o admin)",
        "responses": {
          "202": {
            "description": "Sync encolado (un job chunked por fuente). Estado vía /api/sync/status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "queued": {
                      "type": "boolean"
                    },
                    "jobIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo encolar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}